# docs/swagger.yaml
swagger: '2.0'
info:
  description: API Documentation
  version: '1.1'
  title: Inventory Manager API Documentation
  termsOfService: 'https://www.becommerce.com.br/'
  contact:
    email: rmleme@gmail.com
host: 'localhost:9090'
basePath: /v1/api/inventory
tags:
  - name: inventoryitem
    description: Inventory Manager (Shoes)
paths:
  /items:
    get:
      tags:
        - inventoryitem
      summary: Retrieves all inventory items
      operationId: retrieveAll
      produces:
        - application/json
      parameters:
        - name: shoeId
          in: query
          description: ID of a shoe to be filtered
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/IdentifiedInventoryItem'
        '204':
          description: No Content
      deprecated: false
    post:
      tags:
        - inventoryitem
      summary: Adds a new item to the inventory
      operationId: create
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: inventoryitem
          description: Item object that needs to be added to the inventory
          required: true
          schema:
            $ref: '#/definitions/ExtendedInventoryItem'
      responses:
        '201':
          description: Created
          headers:
            Location:
              type: string
              description: URL of the created inventory item
      deprecated: false
  '/items/{id}':
    get:
      tags:
        - inventoryitem
      summary: Retrieves inventory item by ID
      operationId: retrieve
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of an inventory item to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/IdentifiedInventoryItem'
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - inventoryitem
      summary: Updates an existing inventory item
      operationId: update
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of an inventory item that needs to be updated
          required: true
          type: integer
          format: int64
        - in: body
          name: item
          description: Item object that needs to be updated in the inventory
          required: true
          schema:
            $ref: '#/definitions/InventoryItem'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/IdentifiedInventoryItem'
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - inventoryitem
      summary: Deletes an inventory item
      operationId: delete
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: ID of an inventory item that needs to be deleted
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
      deprecated: false
definitions:
  InventoryItem:
    type: object
    required:
      - description
      - total_quantity
      - reserved_quantity
      - available
    properties:
      description:
        type: string
        maxLength: 64
      total_quantity:
        type: number
        format: double
        minimum: 0
      reserved_quantity:
        type: number
        format: double
        minimum: 0
      available:
        type: boolean
    title: Inventory Item
  ExtendedInventoryItem:
    type: object
    required:
      - shoe_id
    properties:
      shoe_id:
        type: integer
        format: int64
    allOf:
      - $ref: '#/definitions/InventoryItem'
    title: Inventory Item
  IdentifiedInventoryItem:
    type: object
    required:
      - id
    properties:
      id:
        type: integer
        format: int64
    allOf:
      - $ref: '#/definitions/ExtendedInventoryItem'
    title: Inventory Item